move DerivedHookConfig to types.ts
