+++ typescript/sdk/src/hook/EvmHookReader.ts
@@ -39,6 +39,7 @@ import {
   AmountRoutingHookConfig,
   ArbL2ToL1HookConfig,
   CCIPHookConfig,
+  DerivedHookConfig,
   DomainRoutingHookConfig,
   FallbackRoutingHookConfig,
   HookConfig,
@@ -52,8 +53,6 @@ import {
   ProtocolFeeHookConfig,
   RoutingHookConfig,
 } from './types.js';
-
-export type DerivedHookConfig = WithAddress<Exclude<HookConfig, Address>>;
 
 export interface HookReader {
   deriveHookConfig(address: Address): Promise<WithAddress<HookConfig>>;
